Improve docker check in install script
If found that the install script was not running in the `docker:stable` docker container unless I created the directory `~/.docker` manually, which does not exist by default. I thought it would be better than instead of just checking for this directory, to also check if the docker command is available before giving up on trying to install.